# Remove obsolete rules