--- a/tests/Composer/Test/Fixtures/installer/update-allow-list-require-new-replace.test
+++ b/tests/Composer/Test/Fixtures/installer/update-allow-list-require-new-replace.test
@@ -48,7 +48,7 @@ Your requirements could not be resolved to an installable set of packages.
 
   Problem 1
     - current/dep is locked to version 1.0.0 and an update of this package was not requested.
-    - new/pkg 1.0.0 cannot be installed as that would require removing current/dep 1.0.0. new/pkg replaces current/dep and thus cannot coexist with it.
+    - new/pkg[1.0.0] cannot be installed as that would require removing current/dep[1.0.0]. new/pkg replaces current/dep and thus cannot coexist with it.
     - Root composer.json requires new/pkg 1.* -> satisfiable by new/pkg[1.0.0].
 
 Use the option --with-all-dependencies to allow updates and removals for packages currently locked to specific versions.