@Seldaek 2020-01-31
Looks good overall for memory usage but CPU time suffers in some cases. @naderman will investigate.