@naderman 2020-04-07
Performance issue seemed resolved with other changes on 2.0 branch.